chore: refactor commands into their own files
