+++ src/commands/index.js
@@ -1,0 +1,8 @@
+const forgeCommand = require("./forgeCommand");
+const generatorOptionsCommand = require("./generatorOptionsCommand");
+const testGeneratorsCommand = require("./testGeneratorsCommand");
+module.exports = {
+  forgeCommand,
+  generatorOptionsCommand,
+  testGeneratorsCommand,
+};
