+++ src/commands/testGenerators.js
@@ -1,7 +1,7 @@
 const path = require("path");
 const shell = require("shelljs");
 
-const log = require("./log");
+const log = require("../log");
 
 function isJson(str) {
   try {
